Split fd.rs into multiple files.
