+++ src/io/dup.rs
@@ -1,44 +1,11 @@
-//! Functions which operate on file descriptors.
+//! Functions which duplicate file descriptors.
 
 use crate::imp;
 use crate::io::{self, OwnedFd};
 use io_lifetimes::AsFd;
-#[cfg(any(
-    all(linux_raw, feature = "procfs"),
-    all(libc, not(any(target_os = "fuchsia", target_os = "wasi")))
-))]
-use {io_lifetimes::BorrowedFd, std::ffi::CString};
 
 #[cfg(not(target_os = "wasi"))]
 pub use imp::io::DupFlags;
-
-/// `isatty(fd)`—Tests whether a file descriptor refers to a terminal.
-///
-/// # References
-///  - [POSIX]
-///  - [Linux]
-///
-/// [POSIX]: https://pubs.opengroup.org/onlinepubs/9699919799/functions/isatty.html
-/// [Linux]: https://man7.org/linux/man-pages/man3/isatty.3.html
-#[inline]
-pub fn isatty<Fd: AsFd>(fd: &Fd) -> bool {
-    let fd = fd.as_fd();
-    imp::syscalls::isatty(fd)
-}
-
-/// Returns a pair of booleans indicating whether the file descriptor is
-/// readable and/or writeable, respectively.
-///
-/// Unlike [`is_file_read_write`], this correctly detects whether sockets
-/// have been shutdown, partially or completely.
-///
-/// [`is_file_read_write`]: crate::fs::is_file_read_write
-#[cfg(not(target_os = "redox"))]
-#[inline]
-pub fn is_read_write<Fd: AsFd>(fd: &Fd) -> io::Result<(bool, bool)> {
-    let fd = fd.as_fd();
-    imp::syscalls::is_read_write(fd)
-}
 
 /// `dup(fd)`—Creates a new `OwnedFd` instance that shares the same
 /// underlying [file description] as `fd`.
@@ -106,47 +73,4 @@ pub fn dup2<Fd: AsFd>(fd: &Fd, new: &OwnedFd) -> io::Result<()> {
 pub fn dup2_with<Fd: AsFd>(fd: &Fd, new: &OwnedFd, flags: DupFlags) -> io::Result<()> {
     let fd = fd.as_fd();
     imp::syscalls::dup2_with(fd, new, flags)
-}
-
-/// `ttyname_r(fd)`
-///
-/// If `reuse` is non-empty, reuse its buffer to store the result if possible.
-///
-/// # References
-///  - [POSIX]
-///  - [Linux]
-///
-/// [POSIX]: https://pubs.opengroup.org/onlinepubs/9699919799/functions/ttyname.html
-/// [Linux]: https://man7.org/linux/man-pages/man3/ttyname.3.html
-#[cfg(any(
-    all(linux_raw, feature = "procfs"),
-    all(libc, not(any(target_os = "fuchsia", target_os = "wasi")))
-))]
-#[cfg_attr(doc_cfg, doc(cfg(feature = "procfs")))]
-#[inline]
-pub fn ttyname<Fd: AsFd, B: Into<Vec<u8>>>(dirfd: &Fd, reuse: B) -> io::Result<CString> {
-    let dirfd = dirfd.as_fd();
-    _ttyname(dirfd, reuse.into())
-}
-
-#[cfg(any(
-    all(linux_raw, feature = "procfs"),
-    all(libc, not(any(target_os = "fuchsia", target_os = "wasi")))
-))]
-fn _ttyname(dirfd: BorrowedFd<'_>, mut buffer: Vec<u8>) -> io::Result<CString> {
-    // This code would benefit from having a better way to read into
-    // uninitialized memory, but that requires `unsafe`.
-    buffer.clear();
-    buffer.resize(256, 0_u8);
-
-    loop {
-        match imp::syscalls::ttyname(dirfd, &mut buffer) {
-            Err(imp::io::Error::RANGE) => buffer.resize(buffer.len() * 2, 0_u8),
-            Ok(len) => {
-                buffer.resize(len, 0);
-                return Ok(CString::new(buffer).unwrap());
-            }
-            Err(errno) => return Err(errno),
-        }
-    }
 }
