+++ src/io/is_read_write.rs
@@ -1,0 +1,18 @@
+//! Functions which operate on file descriptors.
+
+use crate::imp;
+use crate::io;
+use io_lifetimes::AsFd;
+
+/// Returns a pair of booleans indicating whether the file descriptor is
+/// readable and/or writeable, respectively.
+///
+/// Unlike [`is_file_read_write`], this correctly detects whether sockets
+/// have been shutdown, partially or completely.
+///
+/// [`is_file_read_write`]: crate::fs::is_file_read_write
+#[inline]
+pub fn is_read_write<Fd: AsFd>(fd: &Fd) -> io::Result<(bool, bool)> {
+    let fd = fd.as_fd();
+    imp::syscalls::is_read_write(fd)
+}
